refactor(frontend): rename Title component to TitleText
