+++ Client/src/components/common/TitleText/TitleText.types.ts
