+++ Client/src/components/common/TitleText/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TitleText";
